Organize and tune Snowflake connector
