+++ Cargo.toml
@@ -216,6 +216,7 @@ paste = "1.0.12"
 petgraph = "0.6.0"
 pg-client-config = "0.1.2"
 pin-project-lite = "0.2.16"
+pkcs8 = { version = "0.10.2", features = ["3des"] }
 postgres = "0.19.10"
 postgres-openssl = "0.5.1"
 postgresql_embedded = { version = "0.20.0", features = ["bundled"] }
@@ -247,6 +248,7 @@ rmpv = "1.3.0"
 # Pin `roaring` exactly because the crate only supports the current stable
 # Rust and may start requiring newer compiler features in patch releases.
 roaring = "=0.11.3"
+rsa = { version = "0.9.10", features = ["pkcs5"] }
 rstest = "0.15"
 # Make sure this is the same rustls version used by the `tonic` crate.
 # See the `ensure_default_crypto_provider` function.
