+++ crates/snowflake/Cargo.toml
@@ -29,8 +29,11 @@ arrow = { workspace = true, features = ["ipc"] }
 futures-util = { workspace = true }
 base64 = { workspace = true }
 jsonwebtoken = { workspace = true }
-openssl = { workspace = true }
+pkcs8 = { workspace = true }
+rsa = { workspace = true }
+sha2 = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
 
 [dev-dependencies]
 flate2 = { workspace = true }
+rand = { workspace = true }
